feat: :sparkles: added ability to load more user activities
